### 2020-07-22 #2503 @whedon
[PRE REVIEW]: SeroTools: a Python package for Salmonella serotype data analysis
**Submitting author:** @jdbaugher (<a href="http://orcid.org/0000-0003-2466-3047">Joseph D Baugher</a>)
**Repository:** <a href="https://github.com/CFSAN-Biostatistics/SeroTools" target ="_blank">https://github.com/CFSAN-Biostatistics/SeroTools</a>
**Version:** v0.2.0
