### 2020-07-22 #2503 @Kevin-Mattheus-Moerman
@jdbaugher

- [ ] Can you check that missing DOI :point_up:

- [ ] Can you add a dedicated "Statement of need" section (see [our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements))
