### 2020-07-22 #2503 @fboehm
@Kevin-Mattheus-Moerman - in terms of volume, it seems sufficient, with 1400 lines of Python. However, I'm a new editor, so I have less experience with determining scope. What other characteristics do you suggest we examine when assessing scope?

I don't think that I'm familiar enough with existing software to say whether this submission offers new functionality for Python users.
