### 2020-07-22 #2503 @majensen
@Kevin-Mattheus-Moerman I think this work is in scope. I expect there are few FOSS tools that do this particular job, which is an important one in food safety and molecular epidemiology. Because of this, I would bet that the user base within the specialty is significant (@jdbaugher - feel free to chime in here). This software also provides an important service, because it is free to use. Microbial typing analysis programs can be proprietary and too expensive for small labs.
