+++ paper/paper.md
@@ -40,10 +40,9 @@ These additional annotations are captured in the SeroTools repository and employ
 
 ![Standard formatting of the antigenic formula.](antigenic_profile.png)
 
-In recent years, significant technological advances have resulted in a wide range of 
-molecular-based subtyping options, including highly sensitive approaches based on whole 
-genome sequencing (WGS). One such approach involves the application of software tools to 
-WGS data for *in silico* serovar prediction [@SeqSero; @SeqSero2; @shigatyper; @SerotypeFinder; @ECTyper; @hicap], including real-time prediction [@Feng2020]. SeqSero (a *Salmonella*-specific tool) and other *in silico* serovar designation tools have been adopted by U.S. public health agencies as an alternative to serological testing and for quality control applications [@Dowdy2017; @Timme2019]. The advent of new methodologies for serovar determination has engendered a need for method-comparison studies, and has sparked a growing collection of recent publications comparing various laboratory-based and *in silico* serovar predictions [@Cooper2020; @Banerji2020; @Diep2019; @Tang2019; @Ibrahim2018; @Yachison2017; @SeqSero2; @SeqSero]. In light of the growing interest in *in silico* serovar prediction and serotyping method-comparison studies, SeroTools provides tools which fill multiple gaps in the analysis process. It serves as the only multiformat WKL repository accessible for software development. Currently the WKL scheme is available only as a pdf document [@Grimont2007] and as Python lists in SeqSero [@SeqSero] and SeqSero2 [@SeqSero2]. It also provides the only existing tools for querying the WKL scheme, comparing serovars for congruence, and predicting the most abundant serovar for clusters of isolates. Thus, SeroTools addresses multiple critical needs for the efficient analysis of *Salmonella* serotyping within the public health community.
+# Statement of Need
+
+SeroTools addresses multiple critical needs for the efficient analysis of *Salmonella* serotyping data within the public health community. In recent years, significant technological advances have resulted in a wide range of molecular-based subtyping options, including highly sensitive approaches based on whole genome sequencing (WGS). One such approach involves the application of software tools to WGS data for *in silico* serovar prediction [@SeqSero; @SeqSero2; @shigatyper; @SerotypeFinder; @ECTyper; @hicap], including real-time prediction [@Feng2020]. SeqSero (a *Salmonella*-specific tool) and other *in silico* serovar designation tools have been adopted by U.S. public health agencies as an alternative to serological testing and for quality control applications [@Dowdy2017; @Timme2019]. The advent of new methodologies for serovar determination has engendered a need for method-comparison studies, and has sparked a growing collection of recent publications comparing various laboratory-based and *in silico* serovar predictions [@Cooper2020; @Banerji2020; @Diep2019; @Tang2019; @Ibrahim2018; @Yachison2017; @SeqSero2; @SeqSero]. In light of the growing interest in *in silico* serovar prediction and serotyping method-comparison studies, SeroTools provides unique tools which fill multiple gaps in the analysis process. It serves as the only multiformat WKL repository accessible for software development. Currently the WKL scheme is available only as a pdf document [@Grimont2007] and as Python lists in SeqSero [@SeqSero] and SeqSero2 [@SeqSero2]. SeroTools also provides the only existing tools for querying the WKL scheme, comparing serovars for congruence, and predicting the most abundant serovar for clusters of isolates.
 
 # Functionality and Features
 
