+++ paper/paper.bib
@@ -100,7 +100,7 @@
   author={Henriksen, S.D.},
   title={Methods in Microbiology.},
   chapter = {Chapter I: Serotyping of Bacteria.},
-  publisher={Acacdemic Press},
+  publisher={Academic Press},
   editor = {Bergan, T. and Norris, J. R.},
   edition={12th},
   url={https://www.sciencedirect.com/science/article/pii/S0580951708703556#aep-abstract-id7},
