### 2020-07-22 #2503 @jdbaugher
> @jdbaugher
>
>     * [ ]  Can you check that missing DOI ☝️
>
>     * [ ]  Can you add a dedicated "Statement of need" section (see [our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements))

@Kevin-Mattheus-Moerman I have added the Statement of Need.

Regarding, the missing DOI, the reference in question is a book for which I have been unable to locate a DOI. The suggested DOI is for a different publication.
