### 2020-07-22 #2503 @jdbaugher
> @Kevin-Mattheus-Moerman I think this work is in scope. I expect there are few FOSS tools that do this particular job, which is an important one in food safety and molecular epidemiology. Because of this, I would bet that the user base within the specialty is significant (@jdbaugher - feel free to chime in here). This software also provides an important service, because it is free to use. Microbial typing analysis programs can be proprietary and too expensive for small labs.

@Kevin-Mattheus-Moerman @majensen Characterization of Salmonella by serotyping is the standard methodology broadly used by public health and infectious disease communities, thus the potential user base would be substantial, including international public health agencies and research communities in academia and global food manufacturing. SeroTools provides unique functionality for query and comparison of serotyping data which is necessary for computationally efficient analyses as genomic databases continue to grow and computational predictions are replacing laboratory-based methods. To my knowledge there are currently no other software tools which support similar functionality.
