### 2020-07-27 #2503 @majensen
Hi @amoeba - would this [JOSS](https://joss.theoj.org) [submission](#2503) be of interest to you for review? thanks for considering - @majensen
