### 2020-07-29 #2503 @majensen
@Maghnuso welcome to Github and JOSS! I will assign you as a reviewer now.
