### 2020-08-11 #2556 @amoeba
Hey @majensen & @jdbaugher: I sat down to start my review and ran into two things I think it'd be good to address before I get farther in:

1) The submission is _not_ licensed using an OSI-approved license but is licensed with a fairly permission but custom license. @jdbaugher Is my assessment right and is there any chance you have the ability to change the licensing here given the origin of the work?

2) Re: "Substantial scholarly effort", my take is that this submission is right at the breakpoint per [the submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). I see the previous discussion over in https://github.com/openjournals/joss-reviews/issues/2503 which seems to imply we're all good but I wanted to check. @majensen can you give me a 👍  if you think the submission satisfies the "Substantial scholarly effort" criterion specifically? The project, as assessed via the git commit history, appears to be under or near three months of effort and the core of the software ([serotools.py](https://github.com/CFSAN-Biostatistics/SeroTools/blob/master/serotools/serotools.py)) is only 629 lines according to `cloc`.
