### 2020-08-11 CFSAN-Biostatistics/SeroTools#1 @amoeba
JOSS Review
[Following https://github.com/openjournals/joss-reviews/issues/2556...]

Hey @jdbaugher, I've just finished reviewing SeroTools and things look great. You've done a great job here -- from the paper, to the documentation, to the makefile/tests/etc. I've raised some issues that could use @majensen's attention over on https://github.com/openjournals/joss-reviews/issues/2556#issuecomment-671724752. Apologies if you feel like you and the previous editors have already hashed those things out -- I just wanted to make sure we're within the JOSS guidelines.

Beyond those two issues, I only found two things that I think needs your attention.

1) The JOSS guidelines specify a statement of need should be [present in the Documentation](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation). I see you've got a great statement in the paper but the `README` is a bit too specialized for my take on the JOSS guidelines. Do you think you could migrate some of the language from the paper over to the `README` that explains the importance of Salmonella sterotyping and who the target audience is for a slightly more broad audience?
2) In the references, the BibTeX entry `methods-micro` is missing the publisher-provided DOI of https://doi.org/10.1016/S0580-9517(08)70355-6. This should be amended. https://www.sciencedirect.com/science/article/pii/S0580951708703556#aep-abstract-id7 provides a BibTeX export if that helps at all.

Let me know if you have any questions.
