### 2020-08-13 #2556 @majensen
Thanks @amoeba. The license is a good catch- @jdbaugher, this is a hard requirement for JOSS; can you review [Open Source Initiative-approved licenses](https://opensource.org/licenses) to see if you would be willing to apply one of these? If there are special circumstances to discuss, please reach out to me directly (maj \-dot\- fortinbras \-at
- gmail \-dot\- com).

As for scholarly effort, I am satisfied with this work on that score, because lines of code belie the slog through data and the grokking of standard nomenclature to create a means for enabling others to avoid having to do same. That IMO is scholarly work that does not show up in a code base. On the user side, this work has the potential to enable scholarly efforts that may not otherwise be attempted, for example, by labs that don't have the expertise to develop such a tool.
So @amoeba 👍 !
