+++ paper/paper.bib
@@ -97,13 +97,13 @@
 }
 
 @book{methods-micro,
-  author={Henriksen, S.D.},
-  title={Methods in Microbiology.},
+  author = {Henriksen, S.D.},
+  title = {Methods in Microbiology.},
   chapter = {Chapter I: Serotyping of Bacteria.},
-  publisher={Academic Press},
+  publisher = {Academic Press},
   editor = {Bergan, T. and Norris, J. R.},
-  edition={12th},
-  url={https://www.sciencedirect.com/science/article/pii/S0580951708703556#aep-abstract-id7},
+  edition = {12th},
+  doi = {10.1016/S0580-9517(08)70355-6},
   year={1978}
 }
 
