### 2020-08-13 CFSAN-Biostatistics/SeroTools#1 @jdbaugher
@amoeba Thanks for your time and input! I have added the DOI and will work on broadening the README file as indicated.
