### 2020-08-17 #2556 @majensen
@amoeba @jdbaugher I've finally reviewed the [license](https://github.com/CFSAN-Biostatistics/SeroTools/blob/master/LICENSE).

First, I state that I am not an attorney, and the following is merely an interpretation based the question of appropriateness for JOSS publication, and does not represent an opinion on any legal matter.

The SeroTools license as presented is not unusual in government-created open software, in my professional experience -- and also has a accepted precedent in JOSS [here](https://github.com/openjournals/joss-reviews/issues/1144).

However, there is a stronger argument for this, in that the SeroTools license can be construed as a rather less detailed version of an OSI-approved license, the [NASA Open Source Agreement v1.3 (NASA-1.3)](https://opensource.org/licenses/NASA-1.3). This is written in a general format that allows any Government Agency to use it. The SeroTools license, in my view, can be interpreted as NASA-1.3, with FDA as the agency, @jdbaugher as the contact, section 3B using the 3rd alternative paragraph (no copyright), section 3F using the second alternative paragraph (voluntary user registration), and sections 4A and B clearly covering the "no warranty" and "no government endorsement" points in the SeroTools license.

In light of this, I am prepared to go forward with the assumption that the license requirement has been met, and will leave it to an associate editor-in-chief to complain.
