### 2020-08-17 #2556 @amoeba
Thanks for diving into this, @majensen. Your analysis seems reasonable to me so I'll go ahead and check this off. At this point my review is nearly an Accept and just waits on @jdbaugher's last changes Re: https://github.com/CFSAN-Biostatistics/SeroTools/issues/1. Will keep you updated.
