### 2020-08-17 #2556 @Maghnuso
Thanks @amoeba and @majensen for identifying and resolving the potential issues of substantial scholarly effort and whether or not the license requirement has been met. I concur on both as they currently stand. My apologies too for joining the process a little late, planning to evaluate SeroTools today and tomorrow. Best, Maghnus
