### 2020-08-19 CFSAN-Biostatistics/SeroTools#1 @jdbaugher
@amoeba I have enhanced the README file as requested and added a statement of need to the documentation docs/statement_of_need.rst -> <https://serotools.readthedocs.io/en/latest/statement_of_need.html>.  If these changes do not fully satisfy your concerns, I will be happy to modify as necessary.
