### 2020-08-23 #2556 @Maghnuso
Hi @jdbaugher and @majensen, I have reviewed the SeroTools package and manuscript. It's very clearly laid out and as a microbiologist with very little experience using Python or similar packages, I largely found it easy to install and execute. I commend the effort you have made to make this a useful and accessible tool for both computational and non-computational biologists. The paper is well written and SeroTools is clearly a valuable addition for those who wish to analyze serotyping data using the WKL scheme that is also amenable to updates. During my use of the software, I did come across a few small issues that might also be encountered by other biologists like me with limited coding/command line skills:

1. For the 'cluster' command, it was not clear to me that I could not input the commands directly, but rather had to create a .txt file of the tab-delineated commands. When I did this, there was also an error that the input commands were not separated by one tab, so I had to correct that, too (I use a Mac, in case that is part of the issue). Some clear instruction here would be helpful to get novice users past these small barriers.
2. The following links did not work for me (at Docs » Repository):
An Excel spreadsheet (White-Kauffman-LeMinor-Scheme.xlsx)
A tab-delimited text file (White-Kauffman-LeMinor_scheme.tsv)

SeroTools is a very useful software and after a little practice, was easy to use for a biologist with little to no experience in command line prompts. The minimally congruent designation is an important consideration and addition, to the other more intuitive designations that are possible.
