+++ docs/references.rst
@@ -1,3 +1,5 @@
+.. _references-label:
+
 ========
 References
 ========
