### 2020-08-31 #2556 @jdbaugher
Thanks @Maghnuso for your time and expertise! I have corrected the broken documentation hyperlinks at https://serotools.readthedocs.io/en/latest/repository.html. Thanks for catching those! I am not sure how to further clarify the usage instructions for the ``cluster`` command documented at https://serotools.readthedocs.io/en/latest/usage.html#cluster.
