### 2020-09-01 #2556 @majensen
@jdbaugher I think all that is needed here (based on discussions with @Maghnuso ) is connecting the dots a little more. That is, to explicitly say that the data in the code box under the "Input" header is actually _in_ a text file, and then that the text file's name is what goes at <input_file> in the command

`` serotools cluster -i <input_file>``

This seems _almost_ obvious - but we found for microbiologist just getting started in computational analysis, the current doc wasn't quite enough.

I think you could almost do

```
Input File: test.txt

   cluster1    Dunkwa
   cluster1    Dunkwa
   cluster1    Utah
   cluster2    Hull

$ serotools cluster -i test.txt
```
