### 2020-09-01 9b69a0d
Update usage.rst
