### 2020-09-01 b30b6f2
Update usage.rst
