+++ docs/usage.rst
@@ -53,7 +53,7 @@ cluster
 Determine the most abundant serovar(s) for one or more clusters of isolates. Input data must be 
 submitted in the form of a tab-delimited file in which each line consists of a cluster ID and one serovar as follows::
 
-Input - example.txt::
+Input File - example.txt::
 
     cluster1	Dunkwa
     cluster1	Dunkwa
