### 2020-09-01 #2556 @jdbaugher
@majensen @Maghnuso Ok understood. I have updated the Usage documentation with clarification as discussed - https://serotools.readthedocs.io/en/latest/usage.html#cluster
