### 2020-09-01 #2556 @Maghnuso
Thanks @majensen for clarifying my request and @jdbaugher for the changes in usage documentation, they will help the likes of me get full value from the software! All of my boxes are checked and I'm happy to recommend "Accept". -Maghnus
