### 2020-09-02 #2556 @majensen
@jdbaugher we're closing in. I've looked over the paper itself. The text looks fine to me.

Can you consider the following improvements to the bibliography:

* Please format the text in the references so that all Latin binomials are appropriately capitalized/italicized (e.g. _Escherichia coli_, _**S**almonella_), and serovar names (e.g. Lubbock) are capitalized.
* Please format same so that proper names (e.g. White-Kauffmann-Le Minor) and acronyms (e.g. UGA) are capitalized.
* Henriksen ref is pretty long in the tooth. That's ok, but I would maybe include refs to (e.g.) the Pneumo, Salmonella, E. coli chapters in Man Clin Microbiol. The point is, conveying that serotypes are not obsolete, although they may be assessed in new, non-serological ways.

thx
