### 2020-09-02 0fdc795
Update bibliography formatting
