+++ paper/paper.bib
@@ -1,16 +1,16 @@
-@inbook{BAM,
-  author = {Andrews, W. H. and Wang, H. and Jacobson, A. and Hammack, T. S.},
-  title = {Bacteriological Analytical Manual.},
-  chapter = {Chapter 5: Salmonella},
-  publisher = {U.S. Food and Drug Administration},
-  url = {https://www.fda.gov/food/laboratory-methods-food/bacteriological-analytical-manual-bam-chapter-5-salmonella},
+@incollection{BAM,
+  author = {{Andrews}, W. H. and {Wang}, H. and {Jacobson}, A. and {Hammack}, T. S.},
+  booktitle = {{Bacteriological Analytical Manual}},
+  title = {Chapter 5: \textit{Salmonella}},
+  publisher = {{U.S. Food and Drug Administration}},
   year = {2019},
+  url = {https://www.fda.gov/food/laboratory-methods-food/bacteriological-analytical-manual-bam-chapter-5-salmonella},
   urldate = {2020-02-27}
 }
 
 @article{Banerji2020,
-  author = {Banerji, S. and Simon, S. and Tille A. and Fruth, A. and Flieger, A.},
-  title = {Genome-based Salmonella Serotyping as the New Gold Standard},
+  author = {Banerji, S. and Simon, S. and Tille, A. and Fruth, A. and Flieger, A.},
+  title = {Genome-based \emph{{S}almonella} serotyping as the new gold standard},
   journal = {Sci Rep},
   volume = {10},
   number = {1},
@@ -20,8 +20,8 @@
 }
 
 @article{Bugarel2015,
-  author = {Bugarel, M. and den Bakker, H. C. and Nightingale, K. K. and Brichta-Harhay, D. M. and Edrington, T. S. and Loneragan, G. H.},
-  title = {Two Draft Genome Sequences of a New Serovar of Salmonella enterica, Serovar Lubbock},
+  author = {Bugarel, M. and {den Bakker}, H. C. and Nightingale, K. K. and {Brichta-Harhay}, D. M. and Edrington, T. S. and Loneragan, G. H.},
+  title = {Two draft genome sequences of a new serovar of \emph{{S}almonella enterica}, serovar {Lubbock}},
   journal = {Genome Announc},
   volume = {3},
   number = {2},
@@ -30,8 +30,8 @@
 }
 
 @misc{CDC2015,
-  author  = {CDC},
-  title   = {Serotypes and the Importance of Serotyping Salmonella},
+  author  = {{CDC}},
+  title   = {Serotypes and the Importance of Serotyping \emph{{S}almonella}},
   publisher = {U.S. Centers for Disease Control and Prevention},
   url     = {https://www.cdc.gov/salmonella/reportspubs/salmonella-atlas/serotyping-importance.html},
   urldate = {2020-02-27}
@@ -39,7 +39,7 @@
 
 @article{Cooper2020,
   author = {Cooper, A. L. and Low, A. J. and Koziol, A. G. and Thomas, M. C. and Leclair, D. and Tamber, S. and Wong, A. and Blais, B. W. and Carrillo, C. D.},
-  title = {Systematic Evaluation of Whole Genome Sequence-Based Predictions of Salmonella Serotype and Antimicrobial Resistance},
+  title = {Systematic evaluation of whole genome sequence-based predictions of \emph{{S}almonella} serotype and antimicrobial resistance},
   journal = {Front Microbiol},
   volume = {11},
   pages = {549},
@@ -49,7 +49,7 @@
 
 @article{Diep2019,
   author = {Diep, B. and Barretto, C. and Portmann, A. C. and Fournier, C. and Karczmarek, A. and Voets, G. and Li, S. and Deng, X. and Klijn, A.},
[…]
