### 2020-09-02 1183ba9
Update bibliography formatting
