+++ paper/paper.bib
@@ -32,7 +32,7 @@
 @misc{CDC2015,
   author  = {{CDC}},
   title   = {Serotypes and the Importance of Serotyping \emph{{S}almonella}},
-  publisher = {U.S. Centers for Disease Control and Prevention},
+  publisher = {{U.S. Centers for Disease Control and Prevention}},
   url     = {https://www.cdc.gov/salmonella/reportspubs/salmonella-atlas/serotyping-importance.html},
   urldate = {2020-02-27}
 }
