+++ paper/paper.bib
@@ -59,7 +59,7 @@
 
 @article{Dowdy2017,
   author={Dowdy, J},
-  title={Deng Awarded UGA Creative Research Medal for \emph{{S}almonella} Classification Software},
+  title={Deng Awarded {UGA} Creative Research Medal for \emph{{S}almonella} Classification Software},
   url={https://georgia.growingamerica.com/news/2017/05/deng-awarded-uga-creative-research-medal-salmonella-classification-software},
   journal={Growing Georgia},
   year={2017}
@@ -77,8 +77,8 @@
 
 @book{Grimont2007,
   author={Grimont, P. A. and Weill, F. X.},
-  title={Antigenic Formulae of the \emph{{S}almonella} Serovars},
-  publisher={{WHO Collaborating Center for Reference and Research on Salmonella, Institut Pasteur}},
+  title={Antigenic Formulae of the Salmonella Serovars},
+  publisher={{WHO Collaborating Center for Reference and Research on} \emph{{S}almonella}{, Institut Pasteur}}},
   edition={9th},
   address={Paris, France},
   url={https://www.pasteur.fr/sites/default/files/veng_0.pdf},
