+++ paper/paper.bib
@@ -77,7 +77,7 @@
 
 @book{Grimont2007,
   author={Grimont, P. A. and Weill, F. X.},
-  title={Antigenic Formulae of the Salmonella Serovars},
+  title={Antigenic Formulae of the {Salmonella} Serovars},
   publisher={{WHO Collaborating Center for Reference and Research on} \emph{{S}almonella}{, Institut Pasteur}}},
   edition={9th},
   address={Paris, France},
