### 2020-09-02 a2f4eff
Update bibliography formatting
