+++ paper/paper.bib
@@ -159,7 +159,7 @@
 
 @incollection{clin-micro,
   author = {Strockbine, N. A. and Bopp, C. A. and Fields, P. I. and Kaper, J. B. and Nataro, J. P.},
-  booktitle = {Manual of Clinical Microbiology},
+  booktitle = {{Manual of Clinical Microbiology}},
   title = {Chapter 37: \emph{{E}scherichia, {S}higella, and {S}almonella}},
   publisher = {ASM Press},
   address = {Washington, DC},
@@ -179,7 +179,7 @@
 
 @article{Timme2019,
   author = {Timme, R. E. and {Sanchez Leon}, M. and Allard, M. W.},
-  title = {Utilizing the Public GenomeTrakr Database for Foodborne Pathogen Traceback},
+  title = {Utilizing the Public {GenomeTrakr} Database for Foodborne Pathogen Traceback},
   journal = {Methods Mol Biol},
   volume = {1918},
   pages = {201-212},
