### 2020-09-02 ff811d4
Update bibliography
