+++ paper/paper.bib
@@ -146,6 +146,17 @@
   urldate = {2020-02-27}
 }
 
+@incollection{clin-micro36,
+  author = {Ledeboer, N. A. and Doern, G. V.},
+  booktitle = {Manual of Clinical Microbiology},
+  title = {Chapter 36: \emph{{H}aemophilus}},
+  publisher = {ASM Press},
+  address = {Washington, DC},
+  edition = {11th},
+  year = {2015},
+  doi = {10.1128/9781555817381.ch36}
+}
+
 @article{salm-nomen,
   author = {\emph{{S}almonella} {Subcommittee of the Nomenclature Committee of the International Society for Microbiology}},
   title = {The Genus \emph{{S}almonella} Lignières, 1900},
@@ -157,14 +168,26 @@
   year = {1934}
 }
 
-@incollection{clin-micro,
+@incollection{clin-micro22,
+  author = {Spellerberg, B. and Brandt, C.},
+  booktitle = {Manual of Clinical Microbiology},
+  title = {Chapter 22: \emph{{S}treptococcus}},
+  publisher = {ASM Press},
+  address = {Washington, DC},
+  edition = {11th},
+  year = {2015},
+  doi = {10.1128/9781555817381.ch22}
+}
+
+@incollection{clin-micro37,
   author = {Strockbine, N. A. and Bopp, C. A. and Fields, P. I. and Kaper, J. B. and Nataro, J. P.},
   booktitle = {{Manual of Clinical Microbiology}},
   title = {Chapter 37: \emph{{E}scherichia, {S}higella, and {S}almonella}},
   publisher = {ASM Press},
   address = {Washington, DC},
   edition = {11th},
-  year = {2015}
+  year = {2015},
+  doi = {10.1128/9781555817381.ch37}
 }
 
 @article{Tang2019,
