### 2020-09-02 fdcbd1c
Update paper.md
