+++ paper/paper.md
@@ -24,9 +24,9 @@ bibliography: paper.bib
 Subtyping, the ability to differentiate and characterize closely related microorganisms, 
 has historically been a critical component of successful outbreak identification and 
 traceback efforts employed by public health researchers and regulatory agencies for 
-foodborne pathogens. Serological subtyping (or serotyping) has been the standard approach, largely based on antibody binding to surface antigens. The identification of specific antigenic factors has facilitated the creation of serotyping schemes, which define each serovar using a specific (generally unique) combination of antigenic factors. Serotyping schemes have been developed to assist in characterization of many microorganisms [@methods-micro].
+foodborne pathogens. Serological subtyping (or serotyping) has been the standard approach, largely based on antibody binding to surface antigens [@methods-micro]. The identification of specific antigenic factors has facilitated the creation of serotyping schemes, which define each serovar using a specific (generally unique) combination of antigenic factors. Serotyping schemes have been developed to assist in characterization of many microorganisms, including pathogens such as *Salmonella*, *E. coli*, *Shigella* [clin-micro37], *Streptococcus* [clin-micro22], and *H. influenzae* [clin-micro36].
 
-*Salmonella* is a major foodborne pathogen for which serotyping has played a fundamental monitoring role for over 50 years[@CDC2015]. *Salmonella* serotyping is generally based on antibody binding to the O antigen (a surface antigen) and one or more H antigen phases (flagellar antigens) [@clin-micro; @BAM]. The White-Kauffmann-Le Minor (WKL) *Salmonella* scheme specifies the naming and formatting conventions for *Salmonella* serotyping data and the antigenic factors (and other characteristics) which define each serovar [@Grimont2007]. SeroTools includes the 2007 WKL scheme [@Grimont2007] and updates [@supp47; @supp48; @Bugarel2015]. 
+*Salmonella* is a major foodborne pathogen for which serotyping has played a fundamental monitoring role for over 50 years[@CDC2015]. *Salmonella* serotyping is generally based on antibody binding to the O antigen (a surface antigen) and one or more H antigen phases (flagellar antigens) [@clin-micro37; @BAM]. The White-Kauffmann-Le Minor (WKL) *Salmonella* scheme specifies the naming and formatting conventions for *Salmonella* serotyping data and the antigenic factors (and other characteristics) which define each serovar [@Grimont2007]. SeroTools includes the 2007 WKL scheme [@Grimont2007] and updates [@supp47; @supp48; @Bugarel2015]. 
 
 The WKL scheme currently recognizes two species of *Salmonella*,  *S. enterica* and *S. bongori*. *S. enterica* is comprised of six subspecies (subsp.): *enterica* (I), *salamae* (II), *arizonae* (IIIa), *diarizonae* (IIIb), *houtenae* (IV) and *indica* (VI). Note that *S. bongori* is still frequently designated as subsp. V for scheme consistency, although it is no longer considered a subspecies of *S. enterica*. The WKL scheme assigns a unique name (e.g. serovar Enteritidis) to each of the serovars of *S. enterica* subsp. *enterica* (I), while the serovars representing the other subspecies are referred to by their antigenic formulae. The antigenic formula formatting is defined by the WKL scheme and is demonstrated for serovar Agona in Figure 1. The formula contains a subspecies designation and a colon-separated list of antigenic factors for which the following fields are required: O antigen, phase 1 H antigen, and phase 2 H antigen. The field for ‘Other H' antigen includes R phases and  third phases and is present only when populated. An antigenic formula may include additional annotation such as:
 
