+++ paper/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 Subtyping, the ability to differentiate and characterize closely related microorganisms, 
 has historically been a critical component of successful outbreak identification and 
 traceback efforts employed by public health researchers and regulatory agencies for 
-foodborne pathogens. Serological subtyping (or serotyping) has been the standard approach, largely based on antibody binding to surface antigens [@methods-micro]. The identification of specific antigenic factors has facilitated the creation of serotyping schemes, which define each serovar using a specific (generally unique) combination of antigenic factors. Serotyping schemes have been developed to assist in characterization of many microorganisms, including pathogens such as *Salmonella*, *E. coli*, *Shigella* [clin-micro37], *Streptococcus* [clin-micro22], and *H. influenzae* [clin-micro36].
+foodborne pathogens. Serological subtyping (or serotyping) has been the standard approach, largely based on antibody binding to surface antigens [@methods-micro]. The identification of specific antigenic factors has facilitated the creation of serotyping schemes, which define each serovar using a specific (generally unique) combination of antigenic factors. Serotyping schemes have been developed to assist in characterization of many microorganisms, including pathogens such as *Salmonella*, *E. coli*, *Shigella* [@clin-micro37], *Streptococcus* [@clin-micro22], and *H. influenzae* [@clin-micro36].
 
 *Salmonella* is a major foodborne pathogen for which serotyping has played a fundamental monitoring role for over 50 years[@CDC2015]. *Salmonella* serotyping is generally based on antibody binding to the O antigen (a surface antigen) and one or more H antigen phases (flagellar antigens) [@clin-micro37; @BAM]. The White-Kauffmann-Le Minor (WKL) *Salmonella* scheme specifies the naming and formatting conventions for *Salmonella* serotyping data and the antigenic factors (and other characteristics) which define each serovar [@Grimont2007]. SeroTools includes the 2007 WKL scheme [@Grimont2007] and updates [@supp47; @supp48; @Bugarel2015]. 
 
