### 2020-09-02 12039f7
Update bibliography formatting
