+++ paper/paper.bib
@@ -148,7 +148,7 @@
 
 @incollection{clin-micro36,
   author = {Ledeboer, N. A. and Doern, G. V.},
-  booktitle = {Manual of Clinical Microbiology},
+  booktitle = {{Manual of Clinical Microbiology}},
   title = {Chapter 36: \emph{{H}aemophilus}},
   publisher = {ASM Press},
   address = {Washington, DC},
@@ -170,7 +170,7 @@
 
 @incollection{clin-micro22,
   author = {Spellerberg, B. and Brandt, C.},
-  booktitle = {Manual of Clinical Microbiology},
+  booktitle = {{Manual of Clinical Microbiology}},
   title = {Chapter 22: \emph{{S}treptococcus}},
   publisher = {ASM Press},
   address = {Washington, DC},
