### 2020-09-02 4c0205b
Update bibliography formatting
