### 2020-09-03 #2556 @jdbaugher
@majensen I have updated the bibliography to correct (hopefully) all of the issues mentioned (and caught some other conversion issues) and modified the first paragraph as suggested. Definitely looking better! Thanks!
