+++ serotools/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'Joseph D. Baugher, Ph.D.'
 __email__ = 'joseph.baugher@fda.hhs.gov'
-__version__ = '0.2.0'
+__version__ = '0.2.1'
