+++ HISTORY.rst
@@ -3,6 +3,13 @@
 History
 =======
 
+0.2.1 (2020-09-04)
+---------------------
+
+* Updated documentation
+* Added JOSS manuscript
+
+
 0.2.0 (2020-02-17)
 ---------------------
 
