### 2020-09-04 #2556 @majensen
Thanks @jdbaugher - could you do one thing, and name the archive with the name of the paper "SeroTools: a Python package for Salmonella serotype data analysis"? This strongly desired by JOSS.
