### 2020-09-05 #2556 @majensen
@jdbaugher - congrats! This is nice software that should be really useful in the biz. @Maghnuso - you're now officially a bioinformatician, congrats. @amoeba - appreciate your help once again from the Land of the Midnight Sun.
